Fixed data handlers not creating files when saving

+++ src/main/java/host/serenity/serenity/api/file/handler/AbstractJsonDataHandler.java
@@ -26,6 +26,10 @@ public abstract class AbstractJsonDataHandler<T> implements ClientDataHandler {
 
     @Override
     public void save() throws IOException {
+        if (!file.exists() && !file.createNewFile()) {
+            return;
+        }
+        
         BufferedWriter writer = new BufferedWriter(new FileWriter(file));
         writer.write(gson.toJson(getObjectToSave(), targetClass));
         writer.close();
